+++ android/eir/src/main/java/org/smartregister/fhircore/eir/ui/patient/details/PatientDetailsFragment.kt
@@ -54,7 +54,7 @@ import org.smartregister.fhircore.engine.util.extension.extractName
 import org.smartregister.fhircore.engine.util.extension.hide
 import org.smartregister.fhircore.engine.util.extension.show
 
-class PatientDetailsFragment private constructor() : Fragment() {
+class PatientDetailsFragment : Fragment() {
 
   private lateinit var fhirEngine: FhirEngine
 
@@ -111,9 +111,9 @@ class PatientDetailsFragment private constructor() : Fragment() {
       startActivity(
         Intent(requireContext(), RecordVaccineActivity::class.java)
           .putExtras(
-            QuestionnaireActivity.requiredIntentArgs(
+            QuestionnaireActivity.intentArgs(
               clientIdentifier = patientId,
-              form = RECORD_VACCINE_FORM
+              formName = RECORD_VACCINE_FORM
             )
           )
       )
@@ -147,9 +147,9 @@ class PatientDetailsFragment private constructor() : Fragment() {
           startActivity(
             Intent(requireContext(), AdverseEventQuestionnaireActivity::class.java)
               .putExtras(
-                QuestionnaireActivity.requiredIntentArgs(
+                QuestionnaireActivity.intentArgs(
                   clientIdentifier = patientId,
-                  form = ADVERSE_EVENT_FORM,
+                  formName = ADVERSE_EVENT_FORM,
                   immunizationId = immunizationAdverseEventItem.immunizationIds[position]
                 )
               )