+++ android/eir/src/test/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireViewModelTest.kt
@@ -52,12 +52,11 @@ import org.smartregister.fhircore.eir.EirApplication
 import org.smartregister.fhircore.eir.coroutine.CoroutineTestRule
 import org.smartregister.fhircore.eir.robolectric.RobolectricTest
 import org.smartregister.fhircore.eir.shadow.EirApplicationShadow
-import org.smartregister.fhircore.eir.shadow.ShadowNpmPackageProvider
 import org.smartregister.fhircore.eir.shadow.TestUtils
 import org.smartregister.fhircore.engine.data.local.DefaultRepository
 
 /** Created by Ephraim Kigamba - nek.eam@gmail.com on 03-07-2021. */
-@Config(shadows = [EirApplicationShadow::class, ShadowNpmPackageProvider::class])
+@Config(shadows = [EirApplicationShadow::class])
 class QuestionnaireViewModelTest : RobolectricTest() {
 
   private lateinit var fhirEngine: FhirEngine
@@ -194,6 +193,8 @@ class QuestionnaireViewModelTest : RobolectricTest() {
     val questionnaireResponseSlot = slot<QuestionnaireResponse>()
 
     every { questionnaireViewModel.saveBundleResources(any()) } just runs
+    coEvery { questionnaireViewModel.performExtraction(any(), any(), any()) } returns
+      Bundle().apply { addEntry().resource = Patient() }
 
     ReflectionHelpers.setField(context, "workerContextProvider", mockk<SimpleWorkerContext>())
 
@@ -204,9 +205,9 @@ class QuestionnaireViewModelTest : RobolectricTest() {
       questionnaireResponse
     )
 
-    coVerify(exactly = 1) { defaultRepo.save(capture(questionnaireResponseSlot)) }
+    coVerify(exactly = 1, timeout = 2000) { defaultRepo.save(capture(questionnaireResponseSlot)) }
 
-    coVerify(exactly = 1) { questionnaireViewModel.saveBundleResources(any()) }
+    coVerify(exactly = 1, timeout = 2000) { questionnaireViewModel.saveBundleResources(any()) }
 
     Assert.assertEquals(
       "0993ldsfkaljlsnldm",
@@ -239,7 +240,7 @@ class QuestionnaireViewModelTest : RobolectricTest() {
 
     coEvery { questionnaireViewModel.fetchStructureMap(any()) } returns StructureMap()
 
-    runBlocking { structureMapProvider.invoke(resourceUrl) }
+    runBlocking { structureMapProvider.invoke(resourceUrl, SimpleWorkerContext()) }
 
     coVerify { questionnaireViewModel.fetchStructureMap(resourceUrl) }
   }