+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireActivity.kt
@@ -37,16 +37,18 @@ import com.google.android.fhir.logicalId
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.Resource
 import org.hl7.fhir.r4.model.StringType
 import org.smartregister.fhircore.engine.R
+import org.smartregister.fhircore.engine.ui.base.AlertDialogue
 import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showConfirmAlert
-import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showErrorAlert
 import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showProgressAlert
 import org.smartregister.fhircore.engine.ui.base.BaseMultiLanguageActivity
 import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
 import org.smartregister.fhircore.engine.util.DispatcherProvider
 import org.smartregister.fhircore.engine.util.extension.assertIsConfigurable
 import org.smartregister.fhircore.engine.util.extension.createFactory
+import org.smartregister.fhircore.engine.util.extension.find
 import org.smartregister.fhircore.engine.util.extension.showToast
 import timber.log.Timber
 
@@ -71,6 +73,8 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
 
   protected var immunizationId: String? = null
 
+  var readOnly: Boolean = false
+
   private val parser = FhirContext.forR4().newJsonParser()
 
   override fun onCreate(savedInstanceState: Bundle?) {
@@ -88,15 +92,16 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     immunizationId = intent.getStringExtra(ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY)
 
     lifecycleScope.launchWhenCreated {
-      questionnaireViewModel = createViewModel(application)
+      readOnly = intent.getBooleanExtra(QUESTIONNAIRE_READ_ONLY, false)
+      questionnaireViewModel = createViewModel(application, readOnly)
 
-      val form = intent.getStringExtra(QUESTIONNAIRE_ARG_FORM)!!
+      val formName = intent.getStringExtra(QUESTIONNAIRE_ARG_FORM)!!
       // form is either name of form in asset/form-config or questionnaire-id
       // load from assets and get questionnaire or if not found build it from questionnaire
       questionnaireConfig =
-        kotlin.runCatching { questionnaireViewModel.getQuestionnaireConfig(form) }.getOrElse {
+        kotlin.runCatching { questionnaireViewModel.getQuestionnaireConfig(formName) }.getOrElse {
           // load questionnaire from db and build config
-          questionnaire = questionnaireViewModel.loadQuestionnaire(form)!!
+          questionnaire = questionnaireViewModel.loadQuestionnaire(formName)!!
 
           QuestionnaireConfig(
             form = questionnaire.name ?: "",
@@ -114,46 +119,56 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
         title = questionnaireConfig.title
       }
 
+      findViewById<Button>(R.id.btn_save_client_info).apply {
+        setOnClickListener(this@QuestionnaireActivity)
+        if (readOnly) {
+          text = context.getString(R.string.done)
+        }
+      }
+
       // Only add the fragment once, when the activity is first created.
       if (savedInstanceState == null) {
         val fragment =
           FhirCoreQuestionnaireFragment().apply {
-            val parsedQuestionnaire = parser.encodeResourceToString(questionnaire)
+            val questionnaireString = parser.encodeResourceToString(questionnaire)
             arguments =
               when {
-                clientIdentifier == null ->
-                  bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire))
-                clientIdentifier != null -> {
+                clientIdentifier == null -> {
+                  // TODO remove this when SDK bug for validation is fixed
+                  // https://github.com/google/android-fhir/issues/912
 
-                  try {
-                    fhirEngine.load(Patient::class.java, clientIdentifier!!)
-                    val parsedQuestionnaireResponse =
-                      parser.encodeResourceToString(
-                        questionnaireViewModel.generateQuestionnaireResponse(
-                          questionnaire!!,
-                          intent
-                        )
-                      )
-                    bundleOf(
-                      Pair(BUNDLE_KEY_QUESTIONNAIRE, parser.encodeResourceToString(questionnaire)),
-                      Pair(BUNDLE_KEY_QUESTIONNAIRE_RESPONSE, parsedQuestionnaireResponse)
-                    )
-                  } catch (e: ResourceNotFoundException) {
-                    setBarcode(questionnaire, clientIdentifier!!, true)
-                    bundleOf(
-                      Pair(BUNDLE_KEY_QUESTIONNAIRE, parser.encodeResourceToString(questionnaire))
-                    )
+                  // TODO - FIXME - updateFrom does not consider nested structure
+                  // https://github.com/opensrp/fhircore/issues/730
+                  bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, questionnaireString)).apply {
+                    val questionnaireResponse = intent.getStringExtra(QUESTIONNAIRE_RESPONSE)
+                    if (readOnly && questionnaireResponse != null) {
+                      putString(BUNDLE_KEY_QUESTIONNAIRE_RESPONSE, questionnaireResponse)
+                    }
                   }
                 }
-                else -> bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire))
+                clientIdentifier != null -> {
+                  try {
+                    fhirEngine.load(Patient::class.java, clientIdentifier!!)
+                  } catch (e: ResourceNotFoundException) {
+                    setBarcode(questionnaire, clientIdentifier!!, true)
+                  }
+
+                  val parsedQuestionnaireResponse =
+                    parser.encodeResourceToString(
+                      questionnaireViewModel.generateQuestionnaireResponse(questionnaire!!, intent)
+                    )
+                  bundleOf(
+                    Pair(BUNDLE_KEY_QUESTIONNAIRE, parser.encodeResourceToString(questionnaire)),
+                    Pair(BUNDLE_KEY_QUESTIONNAIRE_RESPONSE, parsedQuestionnaireResponse)
+                  )
+                }
+                else -> bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, questionnaireString))
               }
           }
         supportFragmentManager.commit { add(R.id.container, fragment, QUESTIONNAIRE_FRAGMENT_TAG) }
       }
       loadProgress.dismiss()
     }
-
-    findViewById<Button>(R.id.btn_save_client_info).setOnClickListener(this)
   }
 
   private fun setBarcode(questionnaire: Questionnaire, code: String, readonly: Boolean) {
@@ -164,40 +179,25 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     }
   }
 
-  private fun Questionnaire.find(linkId: String): Questionnaire.QuestionnaireItemComponent? {
-    return item.find(linkId, null)
-  }
-  private fun List<Questionnaire.QuestionnaireItemComponent>.find(
-    linkId: String,
-    default: Questionnaire.QuestionnaireItemComponent?
-  ): Questionnaire.QuestionnaireItemComponent? {
-    var result = default
-    forEach {
-      if (it.linkId == linkId) {
-        return it
-      } else if (it.item.isNotEmpty()) {
-        result = it.item.find(linkId, result)
-      }
-    }
-
-    return result
-  }
-
-  open fun createViewModel(application: Application) =
+  open fun createViewModel(application: Application, readOnly: Boolean = false) =
     ViewModelProvider(
       this@QuestionnaireActivity,
-      QuestionnaireViewModel(application).createFactory()
+      QuestionnaireViewModel(application, readOnly).createFactory()
     )[QuestionnaireViewModel::class.java]
 
   override fun onClick(view: View) {
     if (view.id == R.id.btn_save_client_info) {
-      showConfirmAlert(
-        context = this,
-        message = R.string.questionnaire_alert_submit_message,
-        title = R.string.questionnaire_alert_submit_title,
-        confirmButtonListener = { handleQuestionnaireSubmit() },
-        confirmButtonText = R.string.questionnaire_alert_submit_button_title
-      )
+      if (readOnly) {
+        finish()
+      } else {
+        showConfirmAlert(
+          context = this,
+          message = R.string.questionnaire_alert_submit_message,
+          title = R.string.questionnaire_alert_submit_title,
+          confirmButtonListener = { handleQuestionnaireSubmit() },
+          confirmButtonText = R.string.questionnaire_alert_submit_button_title
+        )
+      }
     } else {
       showToast(getString(R.string.error_saving_form))
     }
@@ -219,11 +219,10 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     saveProcessingAlertDialog = showProgressAlert(this, R.string.saving_registration)
 
     val questionnaireResponse = getQuestionnaireResponse()
-
     if (!validQuestionnaireResponse(questionnaireResponse)) {
       saveProcessingAlertDialog.dismiss()
 
-      showErrorAlert(
+      AlertDialogue.showErrorAlert(
         this,
         R.string.questionnaire_alert_invalid_message,
         R.string.questionnaire_alert_invalid_title
@@ -246,15 +245,48 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     )
   }
 
+  // TODO remove this when SDK bug for validation is fixed
+  // https://github.com/google/android-fhir/issues/912
+  fun deepFlat(
+    qItems: List<Questionnaire.QuestionnaireItemComponent>,
+    questionnaireResponse: QuestionnaireResponse,
+    targetQ: MutableList<Questionnaire.QuestionnaireItemComponent>,
+    targetQR: MutableList<QuestionnaireResponse.QuestionnaireResponseItemComponent>,
+  ) {
+    qItems.forEach { qit ->
+      // process each inner item list
+      deepFlat(qit.item, questionnaireResponse, targetQ, targetQR)
+
+      // remove nested structure to prevent validation recursion; it is already processed above
+      qit.item.clear()
+
+      // add questionnaire and response pair for each linkid on same index
+      questionnaireResponse.find(qit.linkId)?.let { qrit ->
+        targetQ.add(qit)
+        targetQR.add(qrit)
+      }
+    }
+  }
+
+  // TODO change this when SDK bug for validation is fixed
+  // https://github.com/google/android-fhir/issues/912
   fun validQuestionnaireResponse(questionnaireResponse: QuestionnaireResponse): Boolean {
-    return QuestionnaireResponseValidator.validate(
-        questionnaire.item,
-        questionnaireResponse.item,
-        this
-      )
-      .values
-      .flatten()
-      .all { it.isValid }
+    // clone questionnaire and response for processing and changing structure
+    val q = parser.parseResource(parser.encodeResourceToString(questionnaire)) as Questionnaire
+    val qr =
+      parser.parseResource(parser.encodeResourceToString(questionnaireResponse)) as
+        QuestionnaireResponse
+
+    // flatten and pair all responses temporarily to fix index mapping issue for questionnaire and
+    // questionnaire response
+    val qItems = mutableListOf<Questionnaire.QuestionnaireItemComponent>()
+    val qrItems = mutableListOf<QuestionnaireResponse.QuestionnaireResponseItemComponent>()
+
+    deepFlat(q.item, qr, qItems, qrItems)
+
+    return QuestionnaireResponseValidator.validate(qItems, qrItems, this).values.flatten().all {
+      it.isValid
+    }
   }
 
   open fun handleQuestionnaireResponse(questionnaireResponse: QuestionnaireResponse) {
@@ -272,21 +304,45 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     const val QUESTIONNAIRE_FRAGMENT_TAG = "questionnaire-fragment-tag"
     const val QUESTIONNAIRE_ARG_PATIENT_KEY = "questionnaire_patient_item_id"
     const val ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY = "adverse_event_immunization_item_id"
-    const val QUESTIONNAIRE_ARG_FORM = "questionnaire_form"
+    const val QUESTIONNAIRE_ARG_FORM = "questionnaire-form-name"
+    const val QUESTIONNAIRE_READ_ONLY = "read-only"
+    const val QUESTIONNAIRE_RESPONSE = "questionnaire-response"
     const val FORM_CONFIGURATIONS = "form_configurations.json"
     const val QUESTIONNAIRE_ARG_BARCODE_KEY = "patient-barcode"
     const val WHO_IDENTIFIER_SYSTEM = "WHO-HCID"
 
-    fun requiredIntentArgs(
-      clientIdentifier: String?,
-      form: String,
-      immunizationId: String? = null
+    fun intentArgs(
+      clientIdentifier: String? = null,
+      formName: String,
+      readOnly: Boolean = false,
+      questionnaireResponse: QuestionnaireResponse? = null,
+      immunizationId: String? = null,
+      populationResources: ArrayList<Resource> = ArrayList()
     ) =
       bundleOf(
         Pair(QUESTIONNAIRE_ARG_PATIENT_KEY, clientIdentifier),
-        Pair(QUESTIONNAIRE_ARG_FORM, form),
+        Pair(QUESTIONNAIRE_ARG_FORM, formName),
+        Pair(QUESTIONNAIRE_READ_ONLY, readOnly),
         Pair(ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY, immunizationId)
       )
+        .apply {
+          val jsonParser = FhirContext.forR4().newJsonParser()
+          if (questionnaireResponse != null) {
+            putString(
+              QUESTIONNAIRE_RESPONSE,
+              jsonParser.encodeResourceToString(questionnaireResponse)
+            )
+          }
+
+          val resourcesList = ArrayList<String>()
+          populationResources.forEach { resource ->
+            resourcesList.add(jsonParser.encodeResourceToString(resource))
+          }
+
+          if (resourcesList.isNotEmpty()) {
+            putStringArrayList(QUESTIONNAIRE_POPULATION_RESOURCES, resourcesList)
+          }
+        }
   }
 
   override fun onOptionsItemSelected(item: MenuItem): Boolean {
@@ -300,12 +356,16 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
   }
 
   override fun onBackPressed() {
-    showConfirmAlert(
-      this,
-      R.string.questionnaire_alert_back_pressed_message,
-      R.string.questionnaire_alert_back_pressed_title,
-      { finish() },
-      R.string.questionnaire_alert_back_pressed_button_title
-    )
+    if (readOnly) {
+      finish()
+    } else {
+      showConfirmAlert(
+        this,
+        R.string.questionnaire_alert_back_pressed_message,
+        R.string.questionnaire_alert_back_pressed_title,
+        { finish() },
+        R.string.questionnaire_alert_back_pressed_button_title
+      )
+    }
   }
 }