+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/register/BaseRegisterActivity.kt
@@ -625,9 +625,9 @@ abstract class BaseRegisterActivity :
     startActivity(
       Intent(this, QuestionnaireActivity::class.java)
         .putExtras(
-          QuestionnaireActivity.requiredIntentArgs(
+          QuestionnaireActivity.intentArgs(
             clientIdentifier = clientIdentifier,
-            form = registerViewModel.registerViewConfiguration.value?.registrationForm!!
+            formName = registerViewModel.registerViewConfiguration.value?.registrationForm!!
           )
         )
     )