+++ android/engine/src/test/java/org/smartregister/fhircore/engine/impl/FhirApplication.kt
@@ -35,18 +35,17 @@ import org.hl7.fhir.r4.context.SimpleWorkerContext
 import org.hl7.fhir.r4.model.Identifier
 import org.hl7.fhir.r4.model.Resource
 import org.hl7.fhir.r4.model.ResourceType
-import org.robolectric.annotation.Config
 import org.smartregister.fhircore.engine.auth.AuthCredentials
 import org.smartregister.fhircore.engine.auth.AuthenticationService
 import org.smartregister.fhircore.engine.configuration.app.ApplicationConfiguration
 import org.smartregister.fhircore.engine.configuration.app.ConfigurableApplication
 import org.smartregister.fhircore.engine.configuration.app.applicationConfigurationOf
-import org.smartregister.fhircore.engine.shadow.ShadowNpmPackageProvider
 import org.smartregister.fhircore.engine.shadow.activity.ShadowLoginActivity
 import org.smartregister.fhircore.engine.util.SecureSharedPreference
 
-@Config(shadows = [ShadowNpmPackageProvider::class])
 class FhirApplication : Application(), ConfigurableApplication {
+
+  val fhirEngineImpl = spyk(FhirEngineImpl())
 
   override val syncJob: SyncJob
     get() = spyk(Sync.basicSyncJob(ApplicationProvider.getApplicationContext()))
@@ -57,7 +56,7 @@ class FhirApplication : Application(), ConfigurableApplication {
     get() = spyk(FhirAuthenticationService())
 
   override val fhirEngine: FhirEngine
-    get() = spyk(FhirEngineImpl())
+    get() = fhirEngineImpl
 
   override val secureSharedPreference: SecureSharedPreference by lazy {
     val secureSharedPreferenceSpy =