+++ android/engine/src/test/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireActivityTest.kt
@@ -25,15 +25,17 @@ import android.widget.TextView
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.fragment.app.commitNow
 import androidx.test.core.app.ApplicationProvider
+import ca.uhn.fhir.context.FhirContext
 import com.google.android.fhir.datacapture.QuestionnaireFragment
-import io.mockk.coEvery
 import io.mockk.every
 import io.mockk.mockkObject
 import io.mockk.spyk
 import io.mockk.unmockkObject
 import io.mockk.verify
+import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.Resource
 import org.hl7.fhir.r4.model.StringType
 import org.junit.After
 import org.junit.Assert
@@ -45,6 +47,8 @@ import org.robolectric.Shadows.shadowOf
 import org.robolectric.shadows.ShadowAlertDialog
 import org.robolectric.util.ReflectionHelpers
 import org.smartregister.fhircore.engine.R
+import org.smartregister.fhircore.engine.configuration.app.ConfigurableApplication
+import org.smartregister.fhircore.engine.impl.FhirApplication
 import org.smartregister.fhircore.engine.robolectric.ActivityRobolectricTest
 import org.smartregister.fhircore.engine.rule.CoroutineTestRule
 import org.smartregister.fhircore.engine.ui.questionnaire.QuestionnaireActivity.Companion.QUESTIONNAIRE_FRAGMENT_TAG
@@ -72,8 +76,8 @@ class QuestionnaireActivityTest : ActivityRobolectricTest() {
     context = ApplicationProvider.getApplicationContext()
     questionnaireViewModel = spyk(QuestionnaireViewModel(context))
 
-    coEvery { questionnaireViewModel.loadQuestionnaire("1234567") } returns
-      Questionnaire().apply { id = "1234567" }
+    ((context as ConfigurableApplication).fhirEngine as FhirApplication.FhirEngineImpl)
+      .mockedResourcesStore.add(Questionnaire().apply { id = "1234567" })
 
     intent =
       Intent().apply {
@@ -105,12 +109,82 @@ class QuestionnaireActivityTest : ActivityRobolectricTest() {
   }
 
   @Test
-  fun testRequiredIntentShouldInsertValues() {
-    val result = QuestionnaireActivity.requiredIntentArgs("1234", "my-form")
+  fun testIntentArgsShouldInsertValues() {
+    val questionnaireResponse = QuestionnaireResponse()
+    val patient = Patient().apply { id = "my-patient-id" }
+    val populationResources = ArrayList<Resource>()
+    populationResources.add(patient)
+    val result =
+      QuestionnaireActivity.intentArgs(
+        "1234",
+        "my-form",
+        true,
+        questionnaireResponse,
+        populationResources = populationResources,
+        immunizationId = "2323"
+      )
     Assert.assertEquals("my-form", result.getString(QuestionnaireActivity.QUESTIONNAIRE_ARG_FORM))
     Assert.assertEquals(
       "1234",
       result.getString(QuestionnaireActivity.QUESTIONNAIRE_ARG_PATIENT_KEY)
+    )
+    Assert.assertTrue(result.getBoolean(QuestionnaireActivity.QUESTIONNAIRE_READ_ONLY))
+    Assert.assertEquals(
+      FhirContext.forR4().newJsonParser().encodeResourceToString(questionnaireResponse),
+      result.getString(QuestionnaireActivity.QUESTIONNAIRE_RESPONSE)
+    )
+    Assert.assertEquals(
+      "2323",
+      result.getString(QuestionnaireActivity.ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY)
+    )
+    Assert.assertEquals(
+      FhirContext.forR4().newJsonParser().encodeResourceToString(patient),
+      result.getStringArrayList(QuestionnaireActivity.QUESTIONNAIRE_POPULATION_RESOURCES)?.get(0)
+    )
+  }
+
+  @Test
+  fun testReadOnlyIntentShouldBeReadToReadOnlyFlag() {
+    Assert.assertFalse(questionnaireActivity.readOnly)
+
+    intent =
+      Intent().apply {
+        putExtra(QuestionnaireActivity.QUESTIONNAIRE_TITLE_KEY, "Patient registration")
+        putExtra(QuestionnaireActivity.QUESTIONNAIRE_ARG_FORM, "patient-registration")
+        putExtra(QuestionnaireActivity.QUESTIONNAIRE_READ_ONLY, true)
+      }
+
+    val questionnaireFragment = spyk<QuestionnaireFragment>()
+    every { questionnaireFragment.getQuestionnaireResponse() } returns QuestionnaireResponse()
+
+    val controller = Robolectric.buildActivity(QuestionnaireActivity::class.java, intent)
+    questionnaireActivity = controller.create().resume().get()
+
+    Assert.assertTrue(questionnaireActivity.readOnly)
+  }
+
+  @Test
+  fun testReadOnlyIntentShouldChangeSaveButtonToDone() {
+    intent =
+      Intent().apply {
+        putExtra(QuestionnaireActivity.QUESTIONNAIRE_TITLE_KEY, "Patient registration")
+        putExtra(QuestionnaireActivity.QUESTIONNAIRE_ARG_FORM, "patient-registration")
+        putExtra(QuestionnaireActivity.QUESTIONNAIRE_READ_ONLY, true)
+      }
+
+    val questionnaireFragment = spyk<QuestionnaireFragment>()
+    every { questionnaireFragment.getQuestionnaireResponse() } returns QuestionnaireResponse()
+
+    val controller = Robolectric.buildActivity(QuestionnaireActivity::class.java, intent)
+    questionnaireActivity = controller.create().resume().get()
+    questionnaireActivity.supportFragmentManager.executePendingTransactions()
+    questionnaireActivity.supportFragmentManager.commitNow {
+      add(questionnaireFragment, QUESTIONNAIRE_FRAGMENT_TAG)
+    }
+
+    Assert.assertEquals(
+      "Done",
+      questionnaireActivity.findViewById<Button>(R.id.btn_save_client_info).text
     )
   }